requirements.txt
Django==2.2
django-modeltranslation==0.17.1
django-summernote==0.8.11.6
Pillow==8.2.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
